refactor tokenizer capability modules
